Make task name a required keyword-only argument
This ensures a missing name results in an early fail. Otherwise the
missing name argument results in an missing attribute error when
the task should be converted to a string. This usually happens if
a task failed and a notification should be sent. An error at this
point prevents a notification all together.