Add a task for checking the number of todays backups
The task gets a list of repositories and counts the number of backups that were
started today. It then uses the notification mechanism to send the counts for
each repository in a single message. This change can reduce the spam by the
backup script. Instead of the success message for each script run it is now
possible to send a message only once a day containing more useful information.