Add tags for selecting subsets of tasks
The backup script now accepts one or more --tag options on the command
line. Furthermore each task must specify a list of tags. A task will
only be executed if at least one of the task's tags matches the tags
of the command line. If no tags are given on the CLI all tasks will
be executed.

This feature allows to execute only a subset of tasks which is
usefull to share a single configuration file between those sets
of tasks. E. g. one can use the same backup repository configuration
for backup and check tasks but isn't forced to execute those
tasks together.