Refactor MergingTaskFactory to remove knowledge about config keys
This class knew about the "type" key in the config. This was a level of
abstraction which didn't fit into the remainder of the class. Now it
doesn't have to know about any config key and just calls a merger and
the task factories. The knowledge about the "type" key is now part of
the creation methods where other config structure knowlegde is located,
too.