Split up generic task and individual commands
Removed the class hierarchy with TaskBase at the top. TaskBase is now called Task.
This class containing the generic task data like a name and the tags set as well
as a command member representing the concrete logic.

Removed config value retrieval from the Task class. This is now part of the
ConfigValueInjector class. Added a method to assign all parameters of the
__init__ method to the self parameter to reduce boilerplate code required
for using the injection.