Clean check command
Add test cases for the command and change the class name to reflect it is
a command and not a task. Renamed the field names to be snake case and
therefore more in line with the rest ot the code base.

The class is still a bit too large and contains several responsibilities
like creating the subprocess call, parsing the subprocess output and
formatting the messages.